FtS: Shades Lib
This change adds minimal hard-coded library with some shades
colors and codes, which allows to test and work on the rest
of the project.

There is a basic type that keeps plain LDR RGB color, and a
small extension that also holds shade name.

The way this is intended to work is copy them across calls
as primitives, thus also adding a common null_shade object
that should be used to signify the colorless (or no color
found) state. The way to recognize this value is either by
its name being NO_COLOR_TAG, or missing.

The long term plan is to turn this into an setting loader.