FtS: Demo
This change adds a cli demo for finding the shade.

The demo consists from setting up the real photo color of
face (the wanted color) and the photo color of the area of
skin where foundation is applied. Then it names the found-
ation, and lets the shade finder sort all known shades by
match to the wanted color.

It is notable that in this case, there is no good solution,
because all closest shades have darker luminosity. The
only one shade that is lighter than the target color is
much brighter (per the luminosity difference) and thus
won't be of much help either.

UI and automated interpretation of results are for some
other chapter.a

In addition to the demo.c, there is also shell script
with gcc one liner that compiles the demo.