FtS: Affine Transformation in YIQ Color Space
This change introduces new module that uses affine color
transformation in the YIQ color space in order to project
the *ideal target shade* (TARGET) based off:

 - Owned real world shade (SAMPLE_SOURCE)
 - Target real world shade (SAMPLE_TARGET)
 - Ideal owned shade (SOURCE)

The process is as follows:

 - SAMPLE colors are converted to the YIQ color space
 - A YIQ color sp[ace transformation is obtained that trans-
   forms the source into the target
 - This transformation is then applied to the YIQ equivalent
   of the SOURCE to get the TARGET

The helper get_yiq_trans() is rather short, but it merges
the three possible transformations:

 - Hue: [1, 0, 0; 0, cos(theta), -sin(theta), 0, sin(theta), cos(theta)]
 - Saturation: [1, 0, 0; 0, S, 0; 0, 0, S]
 - Value: [V, 0, 0; 0, V, 0; 0, 0, V]

into a transformation matrix:

  [V  0              0               ]
  [0  VS cos(theta)  -VS sin(theta) ]
  [0  VS sin(theta)  VS cos(theta)   ]

via a simple equation system.

ODDITIES / WARNS

This shan't really work if the SAMPLE_SOURCE has no green
component in it. This can be looked at.

Normalized RGB colors had been introduced specially for this
module. Their usage may need to be externalized.