FtS: Use Affine Transformation instead of the Linear one
In the beginning, projection of the ideal target color has
been done with a linear scaling transformation. In

  03d0a803f0301d4571fd46b53688a6fcdd6c5898

another possibility was introduced, to use affine transform-
ation in the YIQ color space in order to make the ideal
target shade differ from the ideal source shade same way the
sample source and target shades differ.

This change wires in that piece of code in the algorithm in
effect. Estimate of the ideal wanted shade should better
from now on. Nothing else should change. There are things
that can be fixed / improved as pointed out in the quoted
commit's text, but nothing that should get in the way for
reasonable uses of this lib.