FtS: ANSI C - Init and Decl & Code Orders
Let's make this also compile on earlier compilers, for the
sake of the exercise.

This change addresses the order in which things are done in-
side the functions

  warning: initializer element is not computable at load time [-Wpedantic].

For those occurrences, a variable was first declared and
then populated with value(s).

  error: ‘for’ loop initial declarations are only allowed in C99 or C11 mode
  ISO C90 forbids mixed declarations and code

For those occurrences, variables were declared at the start
of a function.