FtS: More Order around a Search Result
Initially, a Search Result was allocated statically and re-
turned as value, that contains one dynamically allocated
array. The requestor of search result had to clean it up by
free()-ing the array.

This was leaving much space for other requestors to miss how
to release this search results properly, and was costing a
bit to pass them around as they are a bit heavier than the
primitives.

The first change is that the space for entire search result
is allocated dynamically while generated, and pointer to
that memory is returned.

The second change is that a custom free()-er was added, that
should always be used to properly clean up a search result.