FtS: Edge Cases - Memory Freeing
This change alters where the memory allocated for the search
results is freed, as there may be edge cases with corrupted
data that will prove to be leaky.

Consider the search_results being true, but for some reason
having no result_arr pointer. This is still dynamically al-
located memory, but it wouldn't have been freed. The change
moves the freeing outside the processing block and puts it
in its own block, that is entered whenever search_results
resolves to true, assuming that there is memory to free be-
hind anything different than NULL.

The previous code worked well prior to

  0ae42a873d7bcb3efa33bf7ad41087f8c9dbe55b

as the only dynamically allocated memory *was* the
result_arr itself.